fixes
Closes #17491.

+++ Formula/double-conversion.rb
@@ -18,16 +18,18 @@ class DoubleConversion < Formula
 
   def install
     mkdir "dc-build" do
-      system "cmake", "..", "-DBUILD_SHARED_LIBS=ON", *std_cmake_args
+      system "cmake", "..", ("-DBUILD_SHARED_LIBS=ON" unless OS.mac?), *std_cmake_args
       system "make", "install"
     end
 
-    # Move lib64/* to lib/ on Linuxbrew
-    lib64 = Pathname.new "#{lib}64"
-    if lib64.directory?
-      system "mkdir -p #{lib}"
-      system "mv #{lib64}/* #{lib}/"
-      rmdir lib64
+    unless OS.mac?
+      # Move lib64/* to lib/ on Linuxbrew
+      lib64 = Pathname.new "#{lib}64"
+      if lib64.directory?
+        mkdir_p lib
+        mv lib64, lib
+        rmdir lib64
+      end
     end
   end
